# apps/backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "完整的 POS 系統後端 API，包含商品管理、分類管理、庫存管理等功能",
        "title": "POS System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@pos.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api",
    "paths": {
...
        "/product-options": {
            "get": {
                "description": "取得所有商品使用過的自訂 option_groups 與 values（去重）。option_groups 是規格類型，例如「顏色」「尺寸」「款式」；variants 是實際可販售組合。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "取得所有商品選項",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductOptionListResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "取得商品列表，支援搜尋、分類、規格值、價格、庫存狀態、上下架狀態篩選。預購商品 quantity=0 時 stock_status=out-of-stock，但 sale_status=active 仍 can_order=true。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "取得商品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "頁碼",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每頁數量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "商品名稱搜尋",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分類 ID",
                        "name": "categories[]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "規格名稱或規格值搜尋，例如 白色、S、顏色",
                        "name": "options[]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "最低價格",
                        "name": "priceMin",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "最高價格",
                        "name": "priceMax",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "庫存狀態 (all/in-stock/out-of-stock)",
                        "name": "stockStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "上下架狀態 (all/active/inactive)",
                        "name": "saleStatus",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductListResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "新增商品。最多兩層自訂規格；option_groups 為空代表無變體商品，使用商品層級 quantity。若有 option_groups，必須提供 variants，庫存由 variant.quantity 管理。quantity=0 僅代表無現貨，sale_status=active 時仍可預購。\n\n無變體範例：{\"name\":\"銀色項鍊\",\"category_id\":\"...\",\"price\":\"590\",\"quantity\":0,\"sale_status\":\"active\",\"option_groups\":[],\"variants\":[]}\n\n一層規格範例：{\"name\":\"托特包\",\"category_id\":\"...\",\"price\":\"1280\",\"option_groups\":[{\"name\":\"款式\",\"values\":[\"現貨\",\"預購\"]}],\"variants\":[{\"option_values\":{\"款式\":\"現貨\"},\"quantity\":3},{\"option_values\":{\"款式\":\"預購\"},\"quantity\":0}]}\n\n兩層規格範例：{\"name\":\"經典T恤\",\"category_id\":\"...\",\"price\":\"590\",\"option_groups\":[{\"name\":\"顏色\",\"values\":[\"白色\",\"黑色\"]},{\"name\":\"尺寸\",\"values\":[\"S\",\"M\"]}],\"variants\":[{\"option_values\":{\"顏色\":\"白色\",\"尺寸\":\"S\"},\"quantity\":0,\"sale_status\":\"active\"}]}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "新增商品",
                "parameters": [
                    {
                        "description": "商品資訊；無變體商品使用 quantity，有變體商品使用 variants[].quantity",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateProductSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR / BUSINESS_RULE_ERROR / 圖片或分類不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "CONFLICT：重複變體組合",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/batch-delete": {
            "post": {
                "description": "批量刪除多個商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "批量刪除商品",
                "parameters": [
                    {
                        "description": "商品 ID 陣列",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BatchDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/dto.BatchDeleteProductSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR / SOME_PRODUCTS_IN_USE",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "INTERNAL_SERVER_ERROR",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BatchDeleteProductSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.BatchDeleteProductResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.BatchDeleteRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "6d1fb22a-2eb3-4a8a-9750-b4d63b56b971"
                    ]
                }
            }
        },
...
        "dto.CreateProductSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreateProductResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.ProductListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.ProductOptionListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductOptionGroupResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}